# www/public/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "description": "Issued when a browser pairs with the daemon.",
        "in": "cookie",
        "name": "picode_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The HTTP API the PiCode web UI itself speaks — same daemon, same pairing-gated session. Generated from the server's route registration by cmd/picode-openapi; do not edit the JSON by hand. Authenticate with the picode_session cookie issued when you pair a browser (Settings → Pairing).",
    "title": "PiCode HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/apps": {
      "get": {
        "operationId": "getApiApps",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "apps"
        ]
      }
    },
    "/api/apps/{id}/action": {
      "post": {
        "operationId": "postApiAppsAction",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "apps"
        ]
      }
    },
    "/api/apps/{id}/view": {
      "get": {
        "operationId": "getApiAppsView",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "apps"
        ]
      }
    },
    "/api/automations": {
      "get": {
        "operationId": "getApiAutomations",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      },
      "post": {
        "operationId": "postApiAutomations",
        "parameters": null,
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
    "/api/automations/templates": {
      "get": {
        "operationId": "getApiAutomationsTemplates",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
    "/api/automations/{id}": {
      "delete": {
        "operationId": "deleteApiAutomations",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      },
      "get": {
        "operationId": "getApiAutomations",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      },
      "patch": {
        "operationId": "patchApiAutomations",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
    "/api/automations/{id}/fire": {
      "post": {
        "operationId": "postApiAutomationsFire",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
    "/api/automations/{id}/run": {
      "post": {
        "operationId": "postApiAutomationsRun",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
    "/api/automations/{id}/runs": {
      "get": {
        "operationId": "getApiAutomationsRuns",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
    "/api/automations/{id}/secret": {
      "post": {
        "operationId": "postApiAutomationsSecret",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "automations"
        ]
      }
    },
...
    "/api/server": {
      "get": {
        "operationId": "getApiServer",
        "parameters": null,
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Not paired"
          }
        },
        "tags": [
          "server"
        ]
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "servers": [
    {
      "description": "local daemon",
      "url": "http://127.0.0.1:8445"
    }
  ],
...
}